Meta description feature added to page/route handling module
With this update, we have made it easier for users to optimize their pages for
search engines by being able to set the meta description for each page. This is
an important factor in SEO and can greatly improve the visibility and ranking
of a website. The page/route handling module has been updated to allow users to
set the meta description as an argument, and the page will then display the
correct meta description. This will ensure that each page has a unique and
relevant meta description, which can help improve the overall SEO of the
website. We hope that this update will make it easier for users to manage and
optimize their website for search engines.